% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_stgrid.R
\name{dim.stgrid}
\alias{dim.stgrid}
\title{Dimensions of a space-time tessellation}
\usage{
\method{dim}{stgrid}(x)
}
\arguments{
\item{x}{An \code{stgrid} object.}
}
\value{
An integer vector of length 2: \code{c(n_cells, n_bins)}.
}
\description{
Returns the size of the underlying space-time data cube: the number of
spatial cells and the number of temporal bins.
}
\examples{
g <- stgrid(
  bbox      = c(0, 0, 100, 100),
  cell_size = 20,
  t_window  = as.POSIXct(c("2023-01-01", "2023-12-31")),
  t_step    = "1 month"
)
dim(g)
}
