% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{venezuela_2026}
\alias{venezuela_2026}
\title{Venezuela 2026 seismic sequence}
\format{
A data frame with columns:
\describe{
\item{lon}{Longitude (degrees, WGS84)}
\item{lat}{Latitude (degrees, WGS84)}
\item{depth_km}{Depth in kilometres}
\item{magnitude}{Moment magnitude}
\item{time}{Event time (POSIXct, UTC)}
}
}
\source{
\url{https://earthquake.usgs.gov/fdsnws/event/1/}
}
\usage{
venezuela_2026
}
\description{
Earthquakes of magnitude 2.5 or greater within 200 km of the
24 June 2026 Mw 7.5 mainshock near Yumare, Venezuela.
Downloaded from the USGS Earthquake Catalog API.
}
\keyword{datasets}
